refactor: move use fns to utils
