+++ src/modules/utils.js
@@ -1,0 +1,23 @@
+import { useState, useEffect, useRef } from 'react'
+
+export const useDebounce = (value, delay) => {
+    const [debouncedValue, setDebouncedValue] = useState(value)
+    useEffect(() => {
+        const handler = setTimeout(() => {
+            setDebouncedValue(value)
+        }, delay)
+        return () => {
+            clearTimeout(handler)
+        }
+    }, [value, delay])
+    return debouncedValue
+}
+
+export const useDidUpdateEffect = (fn, inputs) => {
+    const didMountRef = useRef(false)
+
+    useEffect(() => {
+        if (didMountRef.current) fn()
+        else didMountRef.current = true
+    }, inputs)
+}
